\name{coord_gran}
\alias{coord_gran}
\docType{data}
\title{
Localisation of stations sampled
}
\description{
\kbd{coord_gran} is a dataframe of 2 observations and 21 variables. 
It corresponds to the localisation of the stations sampled.
}
\usage{data(coord_gran)}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
  }
}

\source{
\cite{Godet, L., Fournier, J., Toupoint, N., Olivier, F. 2009. Mapping and monitoring intertidal benthic habitats: a review of techniques and proposal of a new visual methodology for the European coasts. \emph{Progress in Physical Geography} \strong{33}, 378-402}
}
\references{
\cite{Fournier, J., Godet, L., Bonnot-Courtois, C., Baltzer, A., Caline, B. 2009. Distribution des formations superficielles de l archipel de Chausey (Manche). \emph{Geologie de la France} \strong{1}, 5-17}
}
\examples{
data(coord_gran)
}